modif autorisation

+++ src/view/ArticleView/read.php
@@ -9,6 +9,8 @@
 	<li>Description : <?= $article->getDescription() ?></li>
 	<li>Prix : <?= $article->getPrix() ?>€</li>
 </ul>
-<button class="d-flex align-center"><a class="linkToArticles" href=<?=$href?>>Ajouter au panier</a></button>
-<button class="btn"><a class="linkToArticles" href="index.php?route=article&action=delete&id=<?= $article->getId()?>">Supprimer l'article</a></button>
-<button class="btn"><a class="linkToArticles" href="index.php?route=article&action=modifArt&id=<?= $article->getId()?>">Modifier l'article</a></button>
+<button class="d-flex align-center"><a class="linkToArticles" href=<?= $href ?>>Ajouter au panier</a></button>
+<?php if ($connecter) : ?>
+	<button class="btn"><a class="linkToArticles" href="index.php?route=article&action=delete&id=<?= $article->getId() ?>">Supprimer l'article</a></button>
+	<button class="btn"><a class="linkToArticles" href="index.php?route=article&action=modifArt&id=<?= $article->getId() ?>">Modifier l'article</a></button>
+<?php endif; ?>
